backend/requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
pytesseract
pdfplumber
Pillow
scikit-learn
requests
